changed from ints to decimals

+++ WebService/Models/Requests/RecurringTransactionRequest.cs
@@ -6,7 +6,7 @@ namespace WebService
     {
         public string Category { get; set; }
         public string Description { get; set; }
-        public int Amount { get; set; }
+        public float Amount { get; set; }
         public string FrequencyId { get; set; }
         public string TransactionTypeId { get; set; }
     }
